+++ Core/ComponentBuilder.cs
@@ -1,8 +1,6 @@
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Threading;
 using DBC.ECS;
 using Svelto.Common;
@@ -24,104 +22,6 @@ namespace Svelto.ECS
         {
             return obj.GetEntityComponentType().GetHashCode();
         }
-    }
-
-    public static class BurstCompatibleCounter
-    {
-        public static int counter;        
-    }
-
-    static public class ComponentTypeMap
-    {
-        static readonly FasterDictionary<RefWrapper<Type>, ComponentID> _componentTypeMap = new FasterDictionary<RefWrapper<Type>, ComponentID>();
-        static readonly FasterDictionary<ComponentID, Type> _reverseComponentTypeMap = new FasterDictionary<ComponentID, Type>();
-
-        public static void Add(Type type, ComponentID idData)
-        {
-            _componentTypeMap.Add(type, idData);
-            _reverseComponentTypeMap.Add(idData, type);
-        }
-
-        public static ComponentID FetchID(Type type)
-        {
-            return _componentTypeMap[type];
-        }
-
-        public static Type FetchType(ComponentID id)
-        {
-            return _reverseComponentTypeMap[id];
-        }
-    }
-
-    public class ComponentTypeID<T> where T : struct, _IInternalEntityComponent
-    {
-        static readonly SharedStaticWrapper<ComponentID, ComponentTypeID<T>> _id;
-
-        public static ComponentID id
-        {
-            [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            get => _id.Data;
-        }
-
-        //todo: any reason to not do this? If I don't, I cannot Create filters in ready functions and
-        //I have to remove the CreateFilter method
-        static ComponentTypeID()
-        {
-            Init();
-        }
-
-#if UNITY_BURST     
-        [Unity.Burst.BurstDiscard] 
-        //SharedStatic values must be initialized from not burstified code
-#endif
-        static void Init()
-        {
-            _id.Data = Interlocked.Increment(ref BurstCompatibleCounter.counter);
-            ComponentTypeMap.Add(typeof(T), id);
-        }
-    }
-
-    sealed class ComponentIDDebugProxy
-    {
-        public ComponentIDDebugProxy(ComponentID id)
-        {
-            this._id = id;
-        }
-        
-        public Type type => ComponentTypeMap.FetchType(_id);
-
-        readonly ComponentID _id;
-    }
-    
-    [DebuggerTypeProxy(typeof(ComponentIDDebugProxy))]
-    public struct ComponentID: IEquatable<ComponentID>
-    {
-        public static implicit operator int(ComponentID id)
-        {
-            return id._id;
-        }
-        
-        public static implicit operator uint(ComponentID id)
-        {
-            return (uint)id._id;
-        }
-        
-        public static implicit operator ComponentID(int id)
-        {
-            return new ComponentID() {_id = id};
-        }
-
-        public bool Equals(ComponentID other)
-        {
-            return _id == other._id;
-        }
-
-        public override int GetHashCode()
-        {
-            return _id;
-        }
-        
-        int _id;
     }
 
     public class ComponentBuilder<T> : IComponentBuilder where T : struct, _IInternalEntityComponent